Move relaxedSnapshot intp spec-helper
