+++ plugins/@grouparoo/spec-helper/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./lib/specHelper";
+export * from "./lib/relaxedSnapshot";
